feat: RegisterNotification event
